@@ -20,9 +20,11 @@ tests.
 
 from collections import abc
 import contextlib
+import dataclasses
 import difflib
 import enum
 import errno
+import faulthandler
 import getpass
 import inspect
 import io
@@ -39,46 +41,30 @@ import subprocess
 import sys
 import tempfile
 import textwrap
+import typing
+from typing import Any, AnyStr, BinaryIO, Callable, ContextManager, IO, Iterator, List, Mapping, MutableMapping, MutableSequence, NoReturn, Optional, Sequence, Text, TextIO, Tuple, Type, Union
 import unittest
 from unittest import mock  # pylint: disable=unused-import Allow absltest.mock.
 from urllib import parse
 
-try:
-  # The faulthandler module isn't always available, and pytype doesn't
-  # understand that we're catching ImportError, so suppress the error.
-  # pytype: disable=import-error
-  import faulthandler
-  # pytype: enable=import-error
-except ImportError:
-  # We use faulthandler if it is available.
-  faulthandler = None
-
-from absl import app
+from absl import app  # pylint: disable=g-import-not-at-top
 from absl import flags
 from absl import logging
 from absl.testing import _pretty_print_reporter
 from absl.testing import xml_reporter
 
-# Make typing an optional import to avoid it being a required dependency
-# in Python 2. Type checkers will still understand the imports.
-try:
-  # pylint: disable=unused-import
-  import typing
-  from typing import Any, AnyStr, BinaryIO, Callable, ContextManager, IO, Iterator, List, Mapping, MutableMapping, MutableSequence, Optional, Sequence, Text, TextIO, Tuple, Type, Union
-  # pylint: enable=unused-import
-except ImportError:
-  pass
-else:
-  # Use an if-type-checking block to prevent leakage of type-checking only
-  # symbols. We don't want people relying on these at runtime.
-  if typing.TYPE_CHECKING:
-    # Unbounded TypeVar for general usage
-    _T = typing.TypeVar('_T')
+# Use an if-type-checking block to prevent leakage of type-checking only
+# symbols. We don't want people relying on these at runtime.
+if typing.TYPE_CHECKING:
+  # Unbounded TypeVar for general usage
+  _T = typing.TypeVar('_T')
 
-    import unittest.case
-    _OutcomeType = unittest.case._Outcome  # pytype: disable=module-attr
+  import unittest.case  # pylint: disable=g-import-not-at-top,g-bad-import-order
+
+  _OutcomeType = unittest.case._Outcome  # pytype: disable=module-attr
 
 
+# pylint: enable=g-import-not-at-top
 
 # Re-export a bunch of unittest functions we support so that people don't
 # have to import unittest to get them
@@ -593,8 +579,9 @@ class TestCase(unittest.TestCase):
   longMessage = True
 
   # Exit stacks for per-test and per-class scopes.
-  _exit_stack = None
-  _cls_exit_stack = None
+  if sys.version_info < (3, 11):
+    _exit_stack = None
+    _cls_exit_stack = None
 
   def __init__(self, *args, **kwargs):
     super(TestCase, self).__init__(*args, **kwargs)
@@ -603,17 +590,22 @@ class TestCase(unittest.TestCase):
 
   def setUp(self):
     super(TestCase, self).setUp()
-    # NOTE: Only Python 3 contextlib has ExitStack
-    if hasattr(contextlib, 'ExitStack'):
+    # NOTE: Only Python 3 contextlib has ExitStack and
+    # Python 3.11+ already has enterContext.
+    if hasattr(contextlib, 'ExitStack') and sys.version_info < (3, 11):
       self._exit_stack = contextlib.ExitStack()
       self.addCleanup(self._exit_stack.close)
 
   @classmethod
   def setUpClass(cls):
     super(TestCase, cls).setUpClass()
-    # NOTE: Only Python 3 contextlib has ExitStack and only Python 3.8+ has
-    # addClassCleanup.
-    if hasattr(contextlib, 'ExitStack') and hasattr(cls, 'addClassCleanup'):
+    # NOTE: Only Python 3 contextlib has ExitStack, only Python 3.8+ has
+    # addClassCleanup and Python 3.11+ already has enterClassContext.
+    if (
+        hasattr(contextlib, 'ExitStack')
+        and hasattr(cls, 'addClassCleanup')
+        and sys.version_info < (3, 11)
+    ):
       cls._cls_exit_stack = contextlib.ExitStack()
       cls.addClassCleanup(cls._cls_exit_stack.close)
 
@@ -752,6 +744,9 @@ class TestCase(unittest.TestCase):
     Args:
       manager: The context manager to enter.
     """
+    if sys.version_info >= (3, 11):
+      return self.enterContext(manager)
+
     if not self._exit_stack:
       raise AssertionError(
           'self._exit_stack is not set: enter_context is Py3-only; also make '
@@ -761,6 +756,9 @@ class TestCase(unittest.TestCase):
   @enter_context.classmethod
   def enter_context(cls, manager):  # pylint: disable=no-self-argument
     # type: (ContextManager[_T]) -> _T
+    if sys.version_info >= (3, 11):
+      return cls.enterClassContext(manager)
+
     if not cls._cls_exit_stack:
       raise AssertionError(
           'cls._cls_exit_stack is not set: cls.enter_context requires '
@@ -803,6 +801,7 @@ class TestCase(unittest.TestCase):
   ) -> None:
     """Adds `function` as cleanup when the test case succeeds."""
     outcome = self._outcome
+    assert outcome is not None
     previous_failure_count = (
         len(outcome.result.failures)
         + len(outcome.result.errors)
@@ -822,6 +821,7 @@ class TestCase(unittest.TestCase):
     """Returns whether test is passed. Expected to be called during cleanup."""
     outcome = self._outcome
     if sys.version_info[:2] >= (3, 11):
+      assert outcome is not None
       current_failure_count = (
           len(outcome.result.failures)
           + len(outcome.result.errors)
@@ -930,6 +930,12 @@ class TestCase(unittest.TestCase):
       prefix = [prefix]
       prefix_len = 1
 
+    if isinstance(whole, abc.Mapping) or isinstance(whole, abc.Set):
+      self.fail(
+          'For whole: Mapping or Set objects are not supported, found type: %s'
+          % type(whole),
+          msg,
+      )
     try:
       whole_len = len(whole)
     except (TypeError, NotImplementedError):
@@ -1725,6 +1731,66 @@ class TestCase(unittest.TestCase):
 
     raise self.failureException('\n'.join(message))
 
+  def assertDataclassEqual(self, first, second, msg=None):
+    """Asserts two dataclasses are equal with more informative errors.
+
+    Arguments must both be dataclasses. This compares equality of  individual
+    fields and takes care to not compare fields that are marked as
+    non-comparable. It gives per field differences, which are easier to parse
+    than the comparison of the string representations from assertEqual.
+
+    In cases where the dataclass has a custom __eq__, and it is defined in a
+    way that is inconsistent with equality of comparable fields, we raise an
+    exception without further trying to figure out how they are different.
+
+    Args:
+      first: A dataclass, the first value.
+      second: A dataclass, the second value.
+      msg: An optional str, the associated message.
+
+    Raises:
+      AssertionError: if the dataclasses are not equal.
+    """
+
+    if not dataclasses.is_dataclass(first) or isinstance(first, type):
+      raise self.failureException('First argument is not a dataclass instance.')
+    if not dataclasses.is_dataclass(second) or isinstance(second, type):
+      raise self.failureException(
+          'Second argument is not a dataclass instance.'
+      )
+
+    if first == second:
+      return
+
+    if type(first) is not type(second):
+      self.fail(
+          'Found different dataclass types: %s != %s'
+          % (type(first), type(second)),
+          msg,
+      )
+
+    # Make sure to skip fields that are marked compare=False.
+    different = [
+        (f.name, getattr(first, f.name), getattr(second, f.name))
+        for f in dataclasses.fields(first)
+        if f.compare and getattr(first, f.name) != getattr(second, f.name)
+    ]
+
+    safe_repr = unittest.util.safe_repr  # pytype: disable=module-attr
+    message = ['%s != %s' % (safe_repr(first), safe_repr(second))]
+    if different:
+      message.append('Fields that differ:')
+      message.extend(
+          '%s: %s != %s' % (k, safe_repr(first_v), safe_repr(second_v))
+          for k, first_v, second_v in different
+      )
+    else:
+      message.append(
+          'Cannot detect difference by examining the fields of the dataclass.'
+      )
+
+    raise self.fail('\n'.join(message), msg)
+
   def assertUrlEqual(self, a, b, msg=None):
     """Asserts that urls are equal, ignoring ordering of query params."""
     parsed_a = parse.urlparse(a)
@@ -1765,7 +1831,8 @@ class TestCase(unittest.TestCase):
     # rather than just stopping at the first
     problems = []
 
-    _walk_structure_for_problems(a, b, aname, bname, problems)
+    _walk_structure_for_problems(a, b, aname, bname, problems,
+                                 self.assertEqual, self.failureException)
 
     # Avoid spamming the user toooo much
     if self.maxDiff is not None:
@@ -1818,9 +1885,9 @@ class TestCase(unittest.TestCase):
 
     return super(TestCase, self)._getAssertEqualityFunc(first, second)
 
-  def fail(self, msg=None, prefix=None):
-    """Fail immediately with the given message, optionally prefixed."""
-    return super(TestCase, self).fail(self._formatMessage(prefix, msg))
+  def fail(self, msg=None, user_msg=None) -> NoReturn:
+    """Fail immediately with the given standard message and user message."""
+    return super(TestCase, self).fail(self._formatMessage(user_msg, msg))
 
 
 def _sorted_list_difference(expected, actual):
@@ -1896,7 +1963,9 @@ def _are_both_of_mapping_type(a, b):
       b, abc.Mapping)
 
 
-def _walk_structure_for_problems(a, b, aname, bname, problem_list):
+def _walk_structure_for_problems(
+    a, b, aname, bname, problem_list, leaf_assert_equal_func, failure_exception
+):
   """The recursive comparison behind assertSameStructure."""
   if type(a) != type(b) and not (  # pylint: disable=unidiomatic-typecheck
       _are_both_of_integer_type(a, b) or _are_both_of_sequence_type(a, b) or
@@ -1924,7 +1993,7 @@ def _walk_structure_for_problems(a, b, aname, bname, problem_list):
       if k in b:
         _walk_structure_for_problems(
             a[k], b[k], '%s[%r]' % (aname, k), '%s[%r]' % (bname, k),
-            problem_list)
+            problem_list, leaf_assert_equal_func, failure_exception)
       else:
         problem_list.append(
             "%s has [%r] with value %r but it's missing in %s" %
@@ -1942,7 +2011,7 @@ def _walk_structure_for_problems(a, b, aname, bname, problem_list):
     for i in range(minlen):
       _walk_structure_for_problems(
           a[i], b[i], '%s[%d]' % (aname, i), '%s[%d]' % (bname, i),
-          problem_list)
+          problem_list, leaf_assert_equal_func, failure_exception)
     for i in range(minlen, len(a)):
       problem_list.append('%s has [%i] with value %r but %s does not' %
                           (aname, i, a[i], bname))
@@ -1951,7 +2020,9 @@ def _walk_structure_for_problems(a, b, aname, bname, problem_list):
                           (aname, i, bname, b[i]))
 
   else:
-    if a != b:
+    try:
+      leaf_assert_equal_func(a, b)
+    except failure_exception:
       problem_list.append('%s is %r but %s is %r' % (aname, a, bname, b))
 
 
@@ -2074,7 +2145,7 @@ def _is_in_app_main():
 def _register_sigterm_with_faulthandler():
   # type: () -> None
   """Have faulthandler dump stacks on SIGTERM.  Useful to diagnose timeouts."""
-  if faulthandler and getattr(faulthandler, 'register', None):
+  if getattr(faulthandler, 'register', None):
     # faulthandler.register is not available on Windows.
     # faulthandler.enable() is already called by app.run.
     try:
@@ -2284,7 +2355,7 @@ class TestLoader(unittest.TestLoader):
     for name in dir(testCaseClass):
       if _is_suspicious_attribute(testCaseClass, name):
         raise TypeError(TestLoader._ERROR_MSG % name)
-    names = super(TestLoader, self).getTestCaseNames(testCaseClass)
+    names = list(super(TestLoader, self).getTestCaseNames(testCaseClass))
     if self._randomize_ordering_seed is not None:
       logging.info(
           'Randomizing test order with seed: %d', self._randomize_ordering_seed)
@@ -2307,8 +2378,7 @@ def get_default_xml_output_filename():
         os.path.splitext(os.path.basename(sys.argv[0]))[0] + '.xml')
 
 
-def _setup_filtering(argv):
-  # type: (MutableSequence[Text]) -> None
+def _setup_filtering(argv: MutableSequence[str]) -> bool:
   """Implements the bazel test filtering protocol.
 
   The following environment variable is used in this method:
@@ -2323,16 +2393,20 @@ def _setup_filtering(argv):
 
   Args:
     argv: the argv to mutate in-place.
+
+  Returns:
+    Whether test filtering is requested.
   """
   test_filter = os.environ.get('TESTBRIDGE_TEST_ONLY')
   if argv is None or not test_filter:
-    return
+    return False
 
   filters = shlex.split(test_filter)
   if sys.version_info[:2] >= (3, 7):
     filters = ['-k=' + test_filter for test_filter in filters]
 
   argv[1:1] = filters
+  return True
 
 
 def _setup_test_runner_fail_fast(argv):
@@ -2359,8 +2433,9 @@ def _setup_test_runner_fail_fast(argv):
   argv[1:1] = ['--failfast']
 
 
-def _setup_sharding(custom_loader=None):
-  # type: (Optional[unittest.TestLoader]) -> unittest.TestLoader
+def _setup_sharding(
+    custom_loader: Optional[unittest.TestLoader] = None,
+) -> Tuple[unittest.TestLoader, Optional[int]]:
   """Implements the bazel sharding protocol.
 
   The following environment variables are used in this method:
@@ -2379,8 +2454,10 @@ def _setup_sharding(custom_loader=None):
     custom_loader: A TestLoader to be made sharded.
 
   Returns:
-    The test loader for shard-filtering or the standard test loader, depending
-    on the sharding environment variables.
+    A tuple of ``(test_loader, shard_index)``. ``test_loader`` is for
+    shard-filtering or the standard test loader depending on the sharding
+    environment variables. ``shard_index`` is the shard index, or ``None`` when
+    sharding is not used.
   """
 
   # It may be useful to write the shard file even if the other sharding
@@ -2398,7 +2475,7 @@ def _setup_sharding(custom_loader=None):
   base_loader = custom_loader or TestLoader()
   if 'TEST_TOTAL_SHARDS' not in os.environ:
     # Not using sharding, use the expected test loader.
-    return base_loader
+    return base_loader, None
 
   total_shards = int(os.environ['TEST_TOTAL_SHARDS'])
   shard_index = int(os.environ['TEST_SHARD_INDEX'])
@@ -2427,25 +2504,70 @@ def _setup_sharding(custom_loader=None):
     return [x for x in ordered_names if x in filtered_names]
 
   base_loader.getTestCaseNames = getShardedTestCaseNames
-  return base_loader
+  return base_loader, shard_index
 
 
-# pylint: disable=line-too-long
-def _run_and_get_tests_result(argv, args, kwargs, xml_test_runner_class):
-  # type: (MutableSequence[Text], Sequence[Any], MutableMapping[Text, Any], Type) -> unittest.TestResult
-  # pylint: enable=line-too-long
-  """Same as run_tests, except it returns the result instead of exiting."""
+def _run_and_get_tests_result(
+    argv: MutableSequence[str],
+    args: Sequence[Any],
+    kwargs: MutableMapping[str, Any],
+    xml_test_runner_class: Type[unittest.TextTestRunner],
+) -> Tuple[unittest.TestResult, bool]:
+  """Same as run_tests, but it doesn't exit.
+
+  Args:
+    argv: sys.argv with the command-line flags removed from the front, i.e. the
+      argv with which :func:`app.run()<absl.app.run>` has called
+      ``__main__.main``. It is passed to
+      ``unittest.TestProgram.__init__(argv=)``, which does its own flag parsing.
+      It is ignored if kwargs contains an argv entry.
+    args: Positional arguments passed through to
+      ``unittest.TestProgram.__init__``.
+    kwargs: Keyword arguments passed through to
+      ``unittest.TestProgram.__init__``.
+    xml_test_runner_class: The type of the test runner class.
+
+  Returns:
+    A tuple of ``(test_result, fail_when_no_tests_ran)``.
+    ``fail_when_no_tests_ran`` indicates whether the test should fail when
+    no tests ran.
+  """
 
   # The entry from kwargs overrides argv.
   argv = kwargs.pop('argv', argv)
 
+  if sys.version_info[:2] >= (3, 12):
+    # Python 3.12 unittest changed the behavior from PASS to FAIL in
+    # https://github.com/python/cpython/pull/102051. absltest follows this.
+    fail_when_no_tests_ran = True
+  else:
+    # Historically, absltest and unittest before Python 3.12 passes if no tests
+    # ran.
+    fail_when_no_tests_ran = False
+
   # Set up test filtering if requested in environment.
-  _setup_filtering(argv)
+  if _setup_filtering(argv):
+    # When test filtering is requested, ideally we also want to fail when no
+    # tests ran. However, the test filters are usually done when running bazel.
+    # When you run multiple targets, e.g. `bazel test //my_dir/...
+    # --test_filter=MyTest`, you don't necessarily want individual tests to fail
+    # because no tests match in that particular target.
+    # Due to this use case, we don't fail when test filtering is requested via
+    # the environment variable from bazel.
+    fail_when_no_tests_ran = False
+
   # Set up --failfast as requested in environment
   _setup_test_runner_fail_fast(argv)
 
   # Shard the (default or custom) loader if sharding is turned on.
-  kwargs['testLoader'] = _setup_sharding(kwargs.get('testLoader', None))
+  kwargs['testLoader'], shard_index = _setup_sharding(
+      kwargs.get('testLoader', None)
+  )
+  if shard_index is not None and shard_index > 0:
+    # When sharding is requested, all the shards except the first one shall not
+    # fail when no tests ran. This happens when the shard count is greater than
+    # the test case count.
+    fail_when_no_tests_ran = False
 
   # XML file name is based upon (sorted by priority):
   # --xml_output_file flag, XML_OUTPUT_FILE variable,
@@ -2523,9 +2645,13 @@ def _run_and_get_tests_result(argv, args, kwargs, xml_test_runner_class):
   # on argv, which is sys.argv without the command-line flags.
   kwargs['argv'] = argv
 
+  # Request unittest.TestProgram to not exit. The exit will be handled by
+  # `absltest.run_tests`.
+  kwargs['exit'] = False
+
   try:
     test_program = unittest.TestProgram(*args, **kwargs)
-    return test_program.result
+    return test_program.result, fail_when_no_tests_ran
   finally:
     if xml_buffer:
       try:
@@ -2535,9 +2661,11 @@ def _run_and_get_tests_result(argv, args, kwargs, xml_test_runner_class):
         xml_buffer.close()
 
 
-def run_tests(argv, args, kwargs):  # pylint: disable=line-too-long
-  # type: (MutableSequence[Text], Sequence[Any], MutableMapping[Text, Any]) -> None
-  # pylint: enable=line-too-long
+def run_tests(
+    argv: MutableSequence[Text],
+    args: Sequence[Any],
+    kwargs: MutableMapping[Text, Any],
+) -> None:
   """Executes a set of Python unit tests.
 
   Most users should call absltest.main() instead of run_tests.
@@ -2558,8 +2686,13 @@ def run_tests(argv, args, kwargs):  # pylint: disable=line-too-long
     kwargs: Keyword arguments passed through to
       ``unittest.TestProgram.__init__``.
   """
-  result = _run_and_get_tests_result(
-      argv, args, kwargs, xml_reporter.TextAndXMLTestRunner)
+  result, fail_when_no_tests_ran = _run_and_get_tests_result(
+      argv, args, kwargs, xml_reporter.TextAndXMLTestRunner
+  )
+  if fail_when_no_tests_ran and result.testsRun == 0 and not result.skipped:
+    # Python 3.12 unittest exits with 5 when no tests ran. The exit code 5 comes
+    # from pytest which does the same thing.
+    sys.exit(5)
   sys.exit(not result.wasSuccessful())
 
 
